api/requirements.txt
fastapi
requests
uvicorn
python-dotenv
pydantic
setuptools

langchain
langchain-community
langchain-core
langchain_experimental
langgraph
langchain-huggingface

langchain-anthropic
langchain-openai

numexpr
wikipedia

pymilvus

PyPDF2
beautifulsoup4
python-docx
readability-lxml
lxml_html_clean
sentence-transformers

pymongo
motor
mongo
